feat(Tools): Adds Color Picker Component
Adds a way to change map colors and to choose those colors from the
Districtr color scheme and other open source libraries. You can sort the
colors by their source and attributes. Color schemes are adjusted based
on the unit count of the map to display the best range in the given
scheme.

Closes #41,#42,#43,#44,#45,#46